requirements.txt
streamlit
pandas
gspread
google-auth
beautifulsoup4
requests
